refactor: controller-service

+++ src/main/java/kr/co/khedu/member/controller/MemberDeleteController.java
@@ -20,11 +20,30 @@ import kr.co.khedu.member.service.MemberServiceImpl;
 public class MemberDeleteController extends HttpServlet {
 	private final MemberService mService = new MemberServiceImpl();
 	
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	protected void doPost(HttpServletRequest request, HttpServletResponse response) 
+			throws ServletException, IOException {
 		request.setCharacterEncoding("UTF-8");
 		response.setContentType("text/html; charset=UTF-8");
 		
 		MemberDTO loginMember = (MemberDTO)request.getSession().getAttribute("loginMember");
+		if (loginMember == null) {
+			response.getWriter().write("failed");
+			return;
+		}
+		
+        // 프론트에서 입력 받은 비밀번호
+        String inputPassword = request.getParameter("password");
+        if (inputPassword == null || inputPassword.trim().isEmpty()) {
+            response.getWriter().write("failed");
+            return;
+        }
+        // 현재 비밀번호와 비교
+        if (!inputPassword.equals(loginMember.getPassword())) {
+            // 불일치 시
+            response.getWriter().write("failed");
+            return;
+        }
+        
 		int memberId = loginMember.getMemberId();
 
 		int result = mService.deleteMember(memberId);
